Clear form after one submission and format add class form
