+++ client/src/css/ClassForm.scss
@@ -1,5 +1,6 @@
 .class-form {
     .popup-form {
         background:black !important;
+        size:'tiny'
     }
 }
