+++ client/src/pages/Profile.jsx
@@ -1,17 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { getAttendingSessions } from '../utils/apiWrapper';
-
 import 'semantic-ui-css/semantic.min.css';
 import '../css/Profile.scss';
 import SessionSummary from '../components/SessionSummary';
 import ClassForm from '../components/ClassForm.jsx';
 import ClassCard from '../components/ClassCard';
-
-/**
- * Returns a sample API response to demonstrate a working backend
- * Returns GET_SAMPLE_FAIL upon failure
- */
-
 
 function Profile({user}) {
 
@@ -20,6 +13,7 @@ function Profile({user}) {
 
   useEffect(() => {
     const populateSessionsAndClasses = async () => {
+      console.log(user)
       const resp = await getAttendingSessions(user._id);
       if (!resp.error) {
         console.log(resp.data.result);
@@ -31,28 +25,16 @@ function Profile({user}) {
     populateSessionsAndClasses();
   }, [user._id]);
 
-  // const sessions = [
-  //   {
-  //     creator: 'Aaron Alexander',
-  //     class: 'CS 124',
-  //     location: 'Grainger Engineering Library',
-  //   },
-  //   {
-  //     creator: 'Aaron Alexander',
-  //     class: 'CS 124',
-  //     location: 'Grainger Engineering Library',
-  //   },
-  // ];
     return (
     <>
-      <h2>{user._id}</h2>
+      <h2>Welcome to your Profile Page, {user.firstName}</h2>
+
       <h1>My Sessions</h1>
       <p></p>
       {sessions.map((session, i) => (
-        <SessionSummary session={session} key={i} />
+        <SessionSummary user={user} session={session} key={i} />
       ))}
       
-      {/* {user.classes && <p>{user.classes[0]}</p>} */}
       <h1>My Classes</h1>
       <div className = "classes-display">
           {classes && classes.map((userClass, j) => (
@@ -60,7 +42,7 @@ function Profile({user}) {
         ))} 
       </div>
       <p></p>
-      {/* {user.classes && <ClassCard classCardText={user.classes[0]} />} */}
+      
       <ClassForm
               button={<button className="small ui button" id="add-class-btn">Add class</button>}
               user={user}
